Update freertos demo: remove junk, drive LEDs
